--- a/src/lib/statistics/abstract_cardinality_estimator.hpp
+++ b/src/lib/statistics/abstract_cardinality_estimator.hpp
@@ -26,7 +26,7 @@ class AbstractCardinalityEstimator {
   /**
    * @return the estimated output row count of @param lqp
    */
-  virtual Cardinality estimate_cardinality(const std::shared_ptr<AbstractLQPNode>& lqp) const = 0;
+  virtual Cardinality estimate_cardinality(const std::shared_ptr<const AbstractLQPNode>& lqp) const = 0;
 
   /**
    * For increased cardinality estimation performance: