--- a/src/lib/statistics/cardinality_estimation_cache.hpp
+++ b/src/lib/statistics/cardinality_estimation_cache.hpp
@@ -9,7 +9,7 @@ class CardinalityEstimationCache {
  public:
   std::optional<JoinGraphStatisticsCache> join_graph_statistics_cache;
 
-  using StatisticsByLQP = std::unordered_map<std::shared_ptr<AbstractLQPNode>, std::shared_ptr<TableStatistics>>;
+  using StatisticsByLQP = std::unordered_map<std::shared_ptr<const AbstractLQPNode>, std::shared_ptr<TableStatistics>>;
   std::optional<StatisticsByLQP> statistics_by_lqp;
 };