`AbstractCardinalityEstimator::estimate_cardinality()` requires `const std::shared_ptr<AbstractLQPNode>& lqp` as an argument. As there is no obvious reason to make changes to an LQP while estimating the cardinality of its output, this PR changes the signature to require a shared pointer to a `const` LQP node and adapts the underlying data structures.
This change seems not to have a stable effect on performance (TPC-H, sf 0.01, 60s/query).

```diff
 +--------------------------+------------------------------------------------+------------------------------------------------+
 |  GIT-HASH                | 2905a7817d9d8d62581160431c07ff657e537441-dirty | da616f0adc9d72689dfd3be7b30526ab666d8e27-dirty |
 |  benchmark_mode          | Ordered                                        | Ordered                                        |
 |  build_type              | release                                        | release                                        |
 |  chunk_size              | 65535                                          | 65535                                          |
 |  clients                 | 1                                              | 1                                              |
 |  compiler                | clang 9.0.0                                    | clang 9.0.0                                    |
 |  cores                   | 0                                              | 0                                              |
 |  date                    | 2020-10-02 14:14:31                            | 2020-10-02 14:37:39                            |
 |  encoding                | {'default': {'encoding': 'Dictionary'}}        | {'default': {'encoding': 'Dictionary'}}        |
 |  indexes                 | False                                          | False                                          |
 |  max_duration            | 60000000000                                    | 60000000000                                    |
 |  max_runs                | -1                                             | -1                                             |
 |  scale_factor            | 0.009999999776482582                           | 0.009999999776482582                           |
 |  time_unit               | ns                                             | ns                                             |
 |  use_prepared_statements | False                                          | False                                          |
 |  using_scheduler         | False                                          | False                                          |
 |  verify                  | False                                          | False                                          |
 |  warmup_duration         | 0                                              | 0                                              |
 +--------------------------+------------------------------------------------+------------------------------------------------+

 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Item     || Latency (ms/iter)  | Change || Throughput (iter/s) | Change | p-value |
 |          ||      old |     new |        ||      old |      new |        |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | TPC-H 01 ||     11.2 |    11.0 |   -1%  ||    89.63 |    90.57 |   +1%  |  0.0000 |
 | TPC-H 02 ||      2.4 |     2.4 |   +0%  ||   424.18 |   423.19 |   -0%  |  0.7689 |
 | TPC-H 03 ||      1.9 |     1.9 |   -1%  ||   531.81 |   536.29 |   +1%  |  0.0000 |
 | TPC-H 04 ||      1.2 |     1.2 |   -2%  ||   822.28 |   839.80 |   +2%  |  0.0000 |
 | TPC-H 05 ||      3.1 |     3.1 |   -1%  ||   324.00 |   325.76 |   +1%  |  0.0378 |
-| TPC-H 06 ||      0.3 |     0.3 |   +7%  ||  3041.38 |  2852.73 |   -6%  |  0.0000 |
-| TPC-H 07 ||      4.1 |     4.6 |  +13%  ||   244.28 |   215.72 |  -12%  |  0.0000 |
 | TPC-H 08 ||     25.5 |    26.2 |   +3%  ||    39.23 |    38.19 |   -3%  |  0.0126 |
 | TPC-H 09 ||      5.0 |     5.0 |   -0%  ||   197.86 |   198.79 |   +0%  |  0.2924 |
 | TPC-H 10 ||      2.0 |     2.0 |   -2%  ||   489.27 |   501.08 |   +2%  |  0.0000 |
 | TPC-H 11 ||      1.0 |     1.0 |   +3%  ||   994.54 |   968.36 |   -3%  |  0.0000 |
 | TPC-H 12 ||      1.6 |     1.6 |   -3%  ||   623.91 |   642.45 |   +3%  |  0.0000 |
 | TPC-H 13 ||      4.5 |     4.4 |   -2%  ||   222.66 |   227.21 |   +2%  |  0.0000 |
 | TPC-H 14 ||      0.8 |     0.8 |   +0%  ||  1200.83 |  1195.58 |   -0%  |  0.0000 |
 | TPC-H 15 ||      2.4 |     2.4 |   -0%  ||   410.57 |   412.31 |   +0%  |  0.0000 |
 | TPC-H 16 ||      4.6 |     4.5 |   -1%  ||   218.70 |   220.30 |   +1%  |  0.0000 |
-| TPC-H 17 ||      0.8 |     0.8 |   +6%  ||  1268.01 |  1192.20 |   -6%  |  0.0000 |
 | TPC-H 18 ||      5.9 |     5.9 |   +1%  ||   169.21 |   168.13 |   -1%  |  0.0000 |
-| TPC-H 19 ||     10.6 |    11.1 |   +5%  ||    94.32 |    90.05 |   -5%  |  0.0000 |
+| TPC-H 20 ||      5.6 |     5.2 |   -8%  ||   177.80 |   193.15 |   +9%  |  0.0000 |
 | TPC-H 21 ||     10.7 |    10.2 |   -4%  ||    93.38 |    97.58 |   +4%  |  0.0000 |
 | TPC-H 22 ||      2.9 |     3.0 |   +1%  ||   339.69 |   335.88 |   -1%  |  0.0000 |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Sum      ||    108.1 |   108.7 |   +1%  ||          |          |        |         |
 | Geomean  ||          |         |        ||          |          |   -1%  |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
```

For comparison, the same benchmark with 30s/query:
```diff
 +Configuration Overview----+------------------------------------------------+------------------------------------------------+
 | Parameter                | estimator-asis.json                            | estimator-new.json                             |
 +--------------------------+------------------------------------------------+------------------------------------------------+
 |  GIT-HASH                | 2905a7817d9d8d62581160431c07ff657e537441-dirty | da616f0adc9d72689dfd3be7b30526ab666d8e27-dirty |
 |  benchmark_mode          | Ordered                                        | Ordered                                        |
 |  build_type              | release                                        | release                                        |
 |  chunk_size              | 65535                                          | 65535                                          |
 |  clients                 | 1                                              | 1                                              |
 |  compiler                | clang 9.0.0                                    | clang 9.0.0                                    |
 |  cores                   | 0                                              | 0                                              |
 |  date                    | 2020-10-02 13:31:07                            | 2020-10-02 13:43:01                            |
 |  encoding                | {'default': {'encoding': 'Dictionary'}}        | {'default': {'encoding': 'Dictionary'}}        |
 |  indexes                 | False                                          | False                                          |
 |  max_duration            | 30000000000                                    | 30000000000                                    |
 |  max_runs                | -1                                             | -1                                             |
 |  scale_factor            | 0.009999999776482582                           | 0.009999999776482582                           |
 |  time_unit               | ns                                             | ns                                             |
 |  use_prepared_statements | False                                          | False                                          |
 |  using_scheduler         | False                                          | False                                          |
 |  verify                  | False                                          | False                                          |
 |  warmup_duration         | 0                                              | 0                                              |
 +--------------------------+------------------------------------------------+------------------------------------------------+

 +----------++----------+---------+--------++----------+----------+--------+----------------------+
 | Item     || Latency (ms/iter)  | Change || Throughput (iter/s) | Change |              p-value |
 |          ||      old |     new |        ||      old |      new |        |                      |
 +----------++----------+---------+--------++----------+----------+--------+----------------------+
 | TPC-H 01 ||     11.0 |    10.8 |   -2%  ||    91.11 |    92.94 |   +2%  | (run time too short) |
-| TPC-H 02 ||      2.6 |     2.9 |  +12%  ||   387.92 |   347.02 |  -11%  | (run time too short) |
 | TPC-H 03 ||      1.7 |     1.8 |   +4%  ||   574.52 |   550.95 |   -4%  | (run time too short) |
 | TPC-H 04 ||      1.2 |     1.2 |   -4%  ||   816.01 |   847.88 |   +4%  | (run time too short) |
 | TPC-H 05 ||      3.2 |     3.0 |   -4%  ||   316.00 |   328.84 |   +4%  | (run time too short) |
 | TPC-H 06 ||      0.4 |     0.4 |   +1%  ||  2797.29 |  2784.81 |   -0%  | (run time too short) |
+| TPC-H 07 ||      7.7 |     5.9 |  -23%  ||   130.19 |   168.29 |  +29%  | (run time too short) |
 | TPC-H 08 ||     26.8 |    25.7 |   -4%  ||    37.28 |    38.95 |   +4%  | (run time too short) |
+| TPC-H 09 ||      5.5 |     5.0 |   -9%  ||   182.62 |   200.50 |  +10%  | (run time too short) |
 | TPC-H 10 ||      2.1 |     2.1 |   +0%  ||   476.85 |   474.89 |   -0%  | (run time too short) |
 | TPC-H 11 ||      1.0 |     1.0 |   +0%  ||   974.32 |   970.98 |   -0%  | (run time too short) |
 | TPC-H 12 ||      1.6 |     1.6 |   -1%  ||   617.03 |   621.61 |   +1%  | (run time too short) |
 | TPC-H 13 ||      4.5 |     4.5 |   +0%  ||   222.85 |   222.78 |   -0%  | (run time too short) |
 | TPC-H 14 ||      0.8 |     0.8 |   -1%  ||  1181.64 |  1193.01 |   +1%  | (run time too short) |
 | TPC-H 15 ||      2.6 |     2.5 |   -4%  ||   388.81 |   403.74 |   +4%  | (run time too short) |
 | TPC-H 16 ||      4.6 |     4.7 |   +2%  ||   214.95 |   210.80 |   -2%  | (run time too short) |
 | TPC-H 17 ||      0.9 |     0.8 |   -2%  ||  1162.75 |  1185.67 |   +2%  | (run time too short) |
 | TPC-H 18 ||      5.9 |     5.9 |   +0%  ||   170.02 |   169.99 |   -0%  | (run time too short) |
 | TPC-H 19 ||     10.8 |    10.5 |   -2%  ||    92.94 |    95.17 |   +2%  | (run time too short) |
+| TPC-H 20 ||      5.4 |     5.0 |   -8%  ||   184.45 |   200.28 |   +9%  | (run time too short) |
 | TPC-H 21 ||     11.2 |    11.0 |   -2%  ||    89.14 |    90.83 |   +2%  | (run time too short) |
 | TPC-H 22 ||      2.8 |     2.8 |   +1%  ||   354.49 |   350.52 |   -1%  | (run time too short) |
 +----------++----------+---------+--------++----------+----------+--------+----------------------+
 | Sum      ||    114.2 |   109.9 |   -4%  ||          |          |        |                      |
 | Geomean  ||          |         |        ||          |          |   +2%  |                      |
 +----------++----------+---------+--------++----------+----------+--------+----------------------+
```